# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Apollo Service",
    "description": "Service for searching and enriching leads via the Apollo API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3004"
    }
  ],
  "components": {
    "schemas": {
...
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "emailStatus": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "linkedinUrl": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "headline": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "seniority": {
            "type": "string",
            "nullable": true
          },
          "departments": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "subdepartments": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "functions": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "twitterUrl": {
            "type": "string",
            "nullable": true
          },
          "githubUrl": {
            "type": "string",
            "nullable": true
          },
          "facebookUrl": {
            "type": "string",
            "nullable": true
          },
          "employmentHistory": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "organizationName": {
                  "type": "string"
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "current": {
                  "type": "boolean"
                }
              }
            }
          },
          "organizationName": {
            "type": "string",
            "nullable": true
          },
          "organizationDomain": {
            "type": "string",
            "nullable": true
          },
          "organizationIndustry": {
            "type": "string",
            "nullable": true
          },
          "organizationSize": {
            "type": "string",
            "nullable": true
          },
          "organizationRevenueUsd": {
            "type": "string",
            "nullable": true
          },
          "organizationWebsiteUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationShortDescription": {
            "type": "string",
            "nullable": true
          },
          "organizationSeoDescription": {
            "type": "string",
            "nullable": true
          },
          "organizationLinkedinUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationTwitterUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationFacebookUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationBlogUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationCrunchbaseUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationAngellistUrl": {
            "type": "string",
            "nullable": true
          },
          "organizationFoundedYear": {
            "type": "number",
            "nullable": true
          },
          "organizationPrimaryPhone": {
            "type": "string",
            "nullable": true
          },
          "organizationPubliclyTradedSymbol": {
            "type": "string",
            "nullable": true
          },
          "organizationPubliclyTradedExchange": {
            "type": "string",
            "nullable": true
          },
          "organizationAnnualRevenuePrinted": {
            "type": "string",
            "nullable": true
          },
          "organizationTotalFunding": {
            "type": "string",
            "nullable": true
          },
          "organizationTotalFundingPrinted": {
            "type": "string",
            "nullable": true
          },
          "organizationLatestFundingRoundDate": {
            "type": "string",
            "nullable": true
          },
          "organizationLatestFundingStage": {
            "type": "string",
            "nullable": true
          },
          "organizationFundingEvents": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "investors": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "currency": {
                  "type": "string"
                }
              }
            }
          },
          "organizationCity": {
            "type": "string",
            "nullable": true
          },
          "organizationState": {
            "type": "string",
            "nullable": true
          },
          "organizationCountry": {
            "type": "string",
            "nullable": true
          },
          "organizationStreetAddress": {
            "type": "string",
            "nullable": true
          },
          "organizationPostalCode": {
            "type": "string",
            "nullable": true
          },
          "organizationTechnologyNames": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "organizationCurrentTechnologies": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "uid": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                }
              }
            }
          },
          "organizationKeywords": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "organizationIndustries": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "organizationSecondaryIndustries": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "organizationNumSuborganizations": {
            "type": "number",
            "nullable": true
          },
          "organizationRetailLocationCount": {
            "type": "number",
            "nullable": true
          },
          "organizationAlexaRanking": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "emailStatus",
          "title",
          "linkedinUrl"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "MatchResponse": {
        "type": "object",
        "properties": {
          "enrichmentId": {
            "type": "string",
            "nullable": true
          },
          "person": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Person"
              },
              {
                "nullable": true
              }
            ]
          },
          "cached": {
            "type": "boolean",
            "description": "True if the result was served from cache (no Apollo API call)."
          }
        },
        "required": [
          "enrichmentId",
          "person",
          "cached"
        ]
      },
      "MatchRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "organizationDomain": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "firstName",
          "lastName",
          "organizationDomain"
        ]
      },
      "MatchBulkResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchBulkResult"
            },
            "description": "Results in the same order as the input items array."
          }
        },
        "required": [
          "results"
        ]
      },
      "MatchBulkResult": {
        "type": "object",
        "properties": {
          "enrichmentId": {
            "type": "string",
            "nullable": true
          },
          "person": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Person"
              },
              {
                "nullable": true
              }
            ]
          },
          "cached": {
            "type": "boolean"
          }
        },
        "required": [
          "enrichmentId",
          "person",
          "cached"
        ]
      },
      "MatchBulkRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchBulkItem"
            },
            "minItems": 1,
            "maxItems": 10
          }
        },
        "required": [
          "items"
        ]
      },
      "MatchBulkItem": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "organizationDomain": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "firstName",
          "lastName",
          "organizationDomain"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/match": {
      "post": {
        "summary": "Match a person by name and organization domain via Apollo",
        "description": "Match a single person by firstName + lastName + organizationDomain. Uses 12-month cache. Costs tracked as apollo-person-match-credit (only charged when email is found).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Caller's run ID — used as parentRunId when creating child runs",
              "example": "run-abc-123"
            },
            "required": true,
            "description": "Caller's run ID — used as parentRunId when creating child runs",
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID",
              "example": "brand-1"
            },
            "required": true,
            "description": "Brand ID",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID",
              "example": "campaign-1"
            },
            "required": true,
            "description": "Campaign ID",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking",
              "example": "lead-gen"
            },
            "required": false,
            "description": "Feature slug for tracking",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug for run tracking",
              "example": "fetch-lead"
            },
            "required": false,
            "description": "Workflow slug for run tracking",
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Match result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/match/bulk": {
      "post": {
        "summary": "Bulk match people by name and organization domain via Apollo",
        "description": "Match up to 10 people by firstName + lastName + organizationDomain. Each item is cached independently. A single run covers the whole batch; costs tracked per item (apollo-person-match-credit, only when email found).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Caller's run ID — used as parentRunId when creating child runs",
              "example": "run-abc-123"
            },
            "required": true,
            "description": "Caller's run ID — used as parentRunId when creating child runs",
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID",
              "example": "brand-1"
            },
            "required": true,
            "description": "Brand ID",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID",
              "example": "campaign-1"
            },
            "required": true,
            "description": "Campaign ID",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking",
              "example": "lead-gen"
            },
            "required": false,
            "description": "Feature slug for tracking",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug for run tracking",
              "example": "fetch-lead"
            },
            "required": false,
            "description": "Workflow slug for run tracking",
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchBulkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk match results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchBulkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}